requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain_google_genai==2.0.8
openpyxl
python-docx